Changed From nltk to SBERT
This may or may not be temporary, it depends on whether prof will allow it when she sees it later.
SBERT utilizes another model that is trained on sentence structure and grammar. This generates an embedded vector, which shows how important each word is to the sentence, which can be used for comparing sentences/descriptions. I'm not entirely confident in this approach as I dont have a full understanding of how SBERT works, but im pushing it for now  just in case