Fixed version mismatch
Took like hour+ to find the right combination of versions online for transformers and sentence-transformers to allow for a proper import of SentenceTransformer. For some reason, versions of transformers past 4.47 and sentence-transformers 3.2 completely moved the file with models in it, breaking imports and the devs behind them havent fixed it. The cell that imports and downloads those modules automatically install the proper versions, but you should probably run those 3 commands at the top of that cell (uninstalls and cache purge), then remember to reset the kernel before running the cells or your version might stay 4.51.x and 3.3+